# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Maintenance API",
        "version": "1.0.0",
        "description": "API documentation for Car Maintenance System",
        "contact": {
            "name": "API Support",
            "email": "support@carmaintenance.com"
        }
    },
    "servers": [
        {
        "url": "https://srv830738.hstgr.cloud/api",
        "description": "Production Server"
        },
        {
            "url": "http://localhost:3000/api",
            "description": "Development Server"
     }
    ],
...
    "paths": {
...
        "/cars/{id}": {
            "get": {
                "summary": "Get a car by ID",
                "tags": [ "Cars" ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The car ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CarDetailed"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a car",
                "tags": [ "Cars" ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The car ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a car",
                "tags": [ "Cars" ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The car ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/maintenance/car/{carId}": {
            "get": {
                "summary": "Get maintenance records for a specific car",
                "tags": [ "Maintenance" ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "carId",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "The car ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of maintenance records for the car",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "results": {
                                            "type": "number",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Maintenance"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Car": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "plateNumber": {
                        "type": "string",
                        "example": "ABC1234"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Camry"
                    },
                    "year": {
                        "type": "number",
                        "example": 2020
                    },
                    "color": {
                        "type": "string",
                        "example": "Silver"
                    },
                    "status": {
                        "type": "string",
                        "enum": [ "available", "in_use", "maintenance" ],
                        "example": "available"
                    },
                    "meterReading": {
                        "type": "number",
                        "example": 123456
                    },
                    "lastMeterUpdate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CarDetailed": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "plateNumber": {
                        "type": "string",
                        "example": "ABC1234"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Camry"
                    },
                    "year": {
                        "type": "number",
                        "example": 2020
                    },
                    "color": {
                        "type": "string",
                        "example": "Silver"
                    },
                    "status": {
                        "type": "string",
                        "enum": [ "available", "in_use", "maintenance" ],
                        "example": "available"
                    },
                    "driver": {
                        "$ref": "#/components/schemas/Driver"
                    },
                    "maintenanceHistory": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Maintenance"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CarInput": {
                "type": "object",
                "properties": {
                    "plateNumber": {
                        "type": "string",
                        "example": "ABC1234"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Camry"
                    },
                    "year": {
                        "type": "number",
                        "example": 2020
                    },
                    "color": {
                        "type": "string",
                        "example": "Silver"
                    },
                    "driver": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "status": {
                        "type": "string",
                        "enum": [ "available", "in_use", "maintenance" ],
                        "example": "available"
                    }
                },
                "required": [ "plateNumber", "brand", "model", "year", "color" ]
            },
            "Driver": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c86"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.smith@example.com"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "nationalId": {
                        "type": "string",
                        "example": "ABC123456"
                    },
                    "licenseNumber": {
                        "type": "string",
                        "example": "DL98765"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, Anytown"
                    },
                    "car": {
                        "$ref": "#/components/schemas/Car"
                    },
                    "role": {
                        "type": "string",
                        "enum": ["driver", "admin"],
                        "example": "driver"
                    }
                }
            },
...
            "Maintenance": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c87"
                    },
                    "car": {
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "60d21b4667d0d8992e610c85"
                            },
                            {
                                "$ref": "#/components/schemas/Car"
                            }
                        ]
                    },
                    "driver": {
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "60d21b4667d0d8992e610c86"
                            },
                            {
                                "$ref": "#/components/schemas/Driver"
                            }
                        ]
                    },
                    "subCategories": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "type": "string",
                                    "example": "60d21b4667d0d8992e610c88"
                                },
                                {
                                    "$ref": "#/components/schemas/SubCategory"
                                }
                            ]
                        }
                    },
                    "description": {
                        "type": "string",
                        "example": "Regular maintenance - oil change and air filter replacement"
                    },
                    "const": {
                        "type": "number",
                        "example": 120
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Category": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c89"
                    },
                    "name": {
                        "type": "string",
                        "example": "Engine"
                    },
                    "subCategories": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "60d21b4667d0d8992e610c88"
                        }
                    }
                }
            },
...
            "SubCategory": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c88"
                    },
                    "name": {
                        "type": "string",
                        "example": "Oil Change"
                    },
                    "category": {
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "60d21b4667d0d8992e610c89"
                            },
                            {
                                "$ref": "#/components/schemas/Category"
                            }
                        ]
                    },
                    "cost": {
                        "type": "number",
                        "example": 50,
                        "description": "Standard cost for mechanic service"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    }
                }
            },
...
        }
    }
}